Refactor data-management
