+++ src/integrationTest/java/uk/gov/hmcts/reform/pip/data/management/controllers/TestingSupportApiTest.java
@@ -5,6 +5,7 @@ import com.microsoft.applicationinsights.web.dependencies.apachecommons.io.IOUti
 import io.zonky.test.db.AutoConfigureEmbeddedDatabase;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -41,6 +42,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @ActiveProfiles("integration")
 @AutoConfigureEmbeddedDatabase(type = AutoConfigureEmbeddedDatabase.DatabaseType.POSTGRES)
 @WithMockUser(username = "admin", authorities = {"APPROLE_api.request.admin"})
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class TestingSupportApiTest extends IntegrationTestBase {
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
 
@@ -80,7 +82,7 @@ class TestingSupportApiTest extends IntegrationTestBase {
     private MockMvc mockMvc;
 
     @BeforeAll
-    static void setup() {
+    void setup() {
         OBJECT_MAPPER.findAndRegisterModules();
     }
 
@@ -218,7 +220,7 @@ class TestingSupportApiTest extends IntegrationTestBase {
     }
 
     private MvcResult uploadPublication() throws Exception {
-        try (InputStream is = PublicationTest.class.getClassLoader().getResourceAsStream("data/artefact.json")) {
+        try (InputStream is = this.getClass().getClassLoader().getResourceAsStream("data/artefact.json")) {
             String payload = new String(IOUtils.toByteArray(Objects.requireNonNull(is)));
 
             MockHttpServletRequestBuilder postRequest = MockMvcRequestBuilders.post(PUBLICATION_URL)
