+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/controllers/location/LocationMetadataControllerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pip.data.management.controllers;
+package uk.gov.hmcts.reform.pip.data.management.controllers.location;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
